% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_decrypt.R
\name{pdf_decrypt}
\alias{pdf_decrypt}
\title{Decrypt a password-protected PDF file}
\usage{
pdf_decrypt(pdf_path, output_path, password)
}
\arguments{
\item{pdf_path}{Input PDF file path}

\item{output_path}{Output PDF file path}

\item{password}{Password for the PDF}
}
\value{
(Invisibly) the output path
}
\description{
Decrypt a password-protected PDF file
}
\details{
Uses the qpdf R package.
}
