% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_delete_pages.R
\name{pdf_delete_pages}
\alias{pdf_delete_pages}
\title{Delete pages from a PDF using qpdf}
\usage{
pdf_delete_pages(pdf_path, from_page, to_page, output_path)
}
\arguments{
\item{pdf_path}{Path to input PDF}

\item{from_page}{First page to delete (integer)}

\item{to_page}{Last page to delete (integer)}

\item{output_path}{Path to save the new PDF}
}
\value{
(Invisibly) the output path
}
\description{
Delete pages from a PDF using qpdf
}
\details{
Uses the qpdf R package. No external system dependencies required.
}
