% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_encrypt.R
\name{pdf_encrypt}
\alias{pdf_encrypt}
\title{Encrypt a PDF file with a password}
\usage{
pdf_encrypt(pdf_path, output_path, password)
}
\arguments{
\item{pdf_path}{Input PDF file path}

\item{output_path}{Output PDF file path}

\item{password}{Password to set for the PDF}
}
\value{
(Invisibly) the output path
}
\description{
Encrypt a PDF file with a password
}
\details{
Uses the qpdf R package.
}
