% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_merge.R
\name{pdf_merge}
\alias{pdf_merge}
\title{Merge multiple PDF files into one}
\usage{
pdf_merge(input_paths, output_path)
}
\arguments{
\item{input_paths}{Character vector of input PDF file paths}

\item{output_path}{Output PDF file path}
}
\value{
(Invisibly) the output path
}
\description{
Merge multiple PDF files into one
}
\details{
Uses the qpdf R package.
}
