% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_split.R
\name{pdf_split}
\alias{pdf_split}
\title{Split a PDF into separate files for each page}
\usage{
pdf_split(pdf_path, output_dir)
}
\arguments{
\item{pdf_path}{Input PDF file path}

\item{output_dir}{Directory to save the split PDFs}
}
\value{
Character vector of output file paths
}
\description{
Split a PDF into separate files for each page
}
\details{
Uses the qpdf R package.
}
